+++ README.md
@@ -5,7 +5,8 @@
 ### How to run this project
 
 - First: clone the repo `git clone git@github.com:ammarkandel/Book-Store-CMS.git` and open the folder `cd Book-Store-CMS`.
-- Make sure you have yarn installed localy then run `yarn` and `npm start`.
+- Second: Make sure you have yarn globally in your machine `yarn --version` if not run `npm install --global yarn`.
+- Third: Make sure you in the project path in the terminal then run `yarn` and `npm start`.
 
 ### Author
