+++ src/components/App.js
@@ -1,7 +1,7 @@
 import './App.css';
-import BooksForm from './containers/BooksForm';
-import BooksList from './containers/BooksList';
-import CategoryFilter from './containers/CategoryFilter';
+import BooksForm from '../containers/BooksForm';
+import BooksList from '../containers/BooksList';
+import CategoryFilter from '../containers/CategoryFilter';
 
 const App = () => (
   <>
