+++ src/components/Book.js
@@ -1,16 +1,18 @@
 import PropTypes from 'prop-types';
-import RemoveButton from './containers/RemoveButton';
 
 const Book = (props) => {
-  const { id, category, title } = props;
+  const {
+    id, category, title, removeBook,
+  } = props;
 
   return (
     <>
       <tr>
-        <span>{category}</span>
-        <br />
-        <span>{title}</span>
-        <RemoveButton id={id} />
+        <td>
+          {category}
+          {title}
+          <button type="button" onClick={() => removeBook(id)}>Remove</button>
+        </td>
       </tr>
     </>
   );
@@ -20,6 +22,7 @@ Book.propTypes = {
   title: PropTypes.string.isRequired,
   category: PropTypes.string.isRequired,
   id: PropTypes.number.isRequired,
+  removeBook: PropTypes.func.isRequired,
 };
 
 export default Book;
